Further merge changes. Moved schema.js and its tests to the scripts directory.
