+++ assets/src/scripts/schema.spec.js
